implement write functions

+++ BlissGBA/Memory/GeneralMemory.h
@@ -21,7 +21,7 @@ public:
 	void zero();
 	void writeU8(u32 address, u8 value);
 	void writeU16(u32 address, u16 value);
-	void writeU32(u32 address, u16 value);
+	void writeU32(u32 address, u32 value);
 
 	u8 readU8(u32 address);
 	u16 readU16(u32 address);
